SoftFloatState is now a required non-null parameter.
The original intention behind a nullable state was good, but there are now too many things which require it (especially the abstract specialized implementation details which are accessed through the state).

If a user wishes to make it "optional", then they are free to use method extensions (at a possibly slightly higher performance cost due to not being able to lazily get the default state).

The best solution for a user who doesn't want to constantly add the state variable is probably to use the helper stub APIs available directly inside SoftFloatState.