Renamed Tommunism.SoftFloat.Experiments to Tommunism.SoftFloat.Tests. Application can now be used for testing implementations. Currently only the SlowFloat implementation ported from TestFloat 3e has been tested.
The code is a little messy, but there were several revisions for the test runner. Currently TestRunner2 is the newest and most "preferred" implementation, just because it attempts to separate the test generator and test verifier processes into separate classes. TestRunner was the original implementation and might have bugs; but it is single-threaded and that makes it potentially more useful for debugging.

Still need to perform argument parsing, as the current implementation relies on hard-coded test functions and options.