Added support for general and round-trip formats to FloatingDecimal128.
User-defined precision values are still not supported.

Note that these results are not always perfect matches to the Single.ToString and Double.ToString counterparts (even with the same inputs to both). Partly this is because this implementation will sometimes get shorter results than the builtin types. (It also have slightly more preference towards generating integers with trailing zeroes instead of outputting the exponent value.)